requirements.txt
bs4
azure-functions